Create htprocess_bamstats.sh
This is part of an optional HTProcess optional workflow. It runs Bamstats on bam files to help count the mapping of reads to components of a reference file. Built for the use of those needing to identify microbial sequences in a sequencing file, but could be used for other purposes, too.